+++ src/components/PrintLibraryCardButton/PrintLibraryCardButton.js
@@ -10,6 +10,9 @@ import {
   Icon,
   exportToCsv,
 } from '@folio/stripes/components';
+import { useProfilePicture } from '@folio/stripes/smart-components';
+import { isPatronUser, isStaffUser, base64ToBlob, isAValidUUID } from '../util';
+
 
 const onlyFields = () => [
   {
@@ -40,6 +43,11 @@ const onlyFields = () => [
 
 const PrintLibraryCardButton = ({ user, patronGroup }) => {
   const callout = useCallout();
+  const { personal: { profilePictureLink }, active } = user;
+  const isPatronOrStaffUser = isPatronUser(user) || isStaffUser(user);
+  const disabled = !active || !isPatronOrStaffUser;
+
+  const { profilePictureData } = useProfilePicture({ profilePictureId: profilePictureLink, isProfilePictureAUUID: isAValidUUID(profilePictureLink) });
 
   const updateUserForExport = useCallback(() => {
     const { personal, expirationDate } = user;
@@ -50,29 +58,100 @@ const PrintLibraryCardButton = ({ user, patronGroup }) => {
     return modifiedUser;
   }, [user, patronGroup]);
 
+  const showCallout = () => {
+    callout.sendCallout({
+      message: <FormattedMessage id="ui-users.errors.generic" />,
+      type: 'error',
+    });
+  };
+
   const exportUserDetails = () => {
     const exportOptions = {
       onlyFields,
-      filename: `${user.barcode}`
+      filename: `Patron_${user.barcode}`
     };
     const data = [updateUserForExport()];
 
     try {
       exportToCsv(data, exportOptions);
-    } catch (err) {
-      callout.sendCallout({
-        message: <FormattedMessage id="ui-users.errors.generic" />,
-        type: 'error',
-      });
+    } catch (error) {
+      console.error('Error downloading the csv:', error); // eslint-disable-line no-console
+      showCallout();
     }
   };
 
+  const downloadImage = (blob) => {
+    const profilePictureFileName = `${user.barcode}.jpg`;
+
+    if (navigator.msSaveBlob) { // IE 10+
+      navigator.msSaveBlob(blob, profilePictureFileName);
+    } else {
+      const link = document.createElement('a');
+      if (link.download !== undefined) { // feature detection
+        // Browsers that support HTML5 download attribute
+        const url = URL.createObjectURL(blob);
+        link.setAttribute('href', url);
+        link.setAttribute('data-test-exportpp-link', 'true');
+        link.setAttribute('download', profilePictureFileName);
+        link.style.visibility = 'hidden';
+        document.body.appendChild(link);
+        // prevent file downloading on testing env
+        if (process.env.NODE_ENV !== 'test') {
+          window.setTimeout(() => {
+            link.click();
+            document.body.removeChild(link);
+          }, 50);
+        }// Need to debounce this click event from others (Pane actionMenuItems dropdown)
+      } else {
+        console.error('Failed to trigger download for CSV data'); // eslint-disable-line no-console
+        throw new Error('Failed to trigger download for CSV data');
+      }
+    }
+  };
+
+  const exportUserProfPicFromBase64String = () => {
+    try {
+      const blob = base64ToBlob(profilePictureData, 'image/png');
+      downloadImage(blob);
+    } catch (error) {
+      console.error('Error downloading the image:', error); // eslint-disable-line no-console
+      showCallout();
+    }
+  };
+
+  const exportProfPicFromExtUrl = async () => {
+    try {
+      const response = await fetch(profilePictureLink);
+
+      if (!response.ok) {
+        throw new Error(`Failed to fetch image: ${response.statusText}`);
+      }
+
+      const blob = await response.blob();
+      downloadImage(blob);
+    } catch (error) {
+      console.error('Error downloading the image:', error); // eslint-disable-line no-console
+      showCallout();
+    }
+  };
+
+  const exportUserProfilePicture = () => {
+    if (isAValidUUID(profilePictureLink)) {
+      exportUserProfPicFromBase64String();
+    } else {
+      exportProfPicFromExtUrl();
+    }
+  };
+
+  /* ************************************************************************
+  * handlePrintLibraryCard method should export
+  * 1. user details
+  * 2. user profile picture if profile picture config is enabled for tenant
+  *    and user has one image against his profile
+  ************************************************************************* */
   const handlePrintLibraryCard = () => {
-    // this method should export
-    // 1. user details
-    // 2. TODO: user profile picture if profile picture config is enabled for tenant
-    //    and user has one image against his profile
     exportUserDetails();
+    if (profilePictureLink) exportUserProfilePicture();
   };
 
   return (
@@ -80,6 +159,7 @@ const PrintLibraryCardButton = ({ user, patronGroup }) => {
       data-testid="print-library-card"
       buttonStyle="dropdownItem"
       onClick={handlePrintLibraryCard}
+      disabled={disabled}
     >
       <Icon icon="print">
         <FormattedMessage id="ui-users.printLibraryCard" />